+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Change history of the project. All the feature updates, bug fixes, breaking chan
 ## [0.0.13]
 
 - Feature: More math functions added. (floor/ceil/round/sign/pow/sin/cos/tan/log/log2/log10)
+- Feature: New function `kv` which converts object into array of key value pairs
 
 ## [0.0.12]